+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 All notable changes to this project will be documented in this file.
+### [1.11.0] - 2020-09-21
+[feature] device traffic viewer
+[fix] traffic acl check
+[fix] traffic getting(microseconds)
 ### [1.10.19] - 2020-09-17
 [fix] traffic microseconds getting
 ### [1.10.18] - 2020-09-16